+++ src/axom/sidre/core/Group.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2017-2021, Lawrence Livermore National Security, LLC and
-// other Axom Project Developers. See the top-level COPYRIGHT file for details.
+// other Axom Project Developers. See the top-level LICENSE file for details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
@@ -12,12 +12,13 @@
   #include "conduit_relay_io_hdf5.hpp"
 #endif
 
+#include "axom/core/Path.hpp"
+
 // Sidre headers
 #include "ListCollection.hpp"
 #include "MapCollection.hpp"
 #include "Buffer.hpp"
 #include "DataStore.hpp"
-#include "SidreUtilities.hpp"
 
 namespace axom
 {
@@ -2292,21 +2293,20 @@ Group* Group::walkPath(std::string& path, bool create_groups_in_path)
 {
   Group* group_ptr = this;
 
-  std::string::size_type pos = detail::find_exclusive(path, s_path_delimiter);
-  if(pos != std::string::npos)
+  // Split path into parts
+  std::vector<std::string> path_parts =
+    axom::Path(path, s_path_delimiter).parts();
+
+  if(path_parts.size() > 0)
   {
-    std::vector<std::string> tokens = detail::split(path, s_path_delimiter, pos);
-    std::vector<std::string>::iterator stop = tokens.end() - 1;
+    // Find stopping point (right before last part of path)
+    std::vector<std::string>::const_iterator stop = path_parts.end() - 1;
 
     // Navigate path down to desired Group
-    for(std::vector<std::string>::const_iterator iter = tokens.begin();
+    for(std::vector<std::string>::const_iterator iter = path_parts.begin();
         iter < stop;
         ++iter)
     {
-      SLIC_ASSERT_MSG(iter->size() > 0,
-                      SIDRE_GROUP_LOG_PREPEND << "Empty name in provided path '"
-                                              << path << "'.");
-
       if(group_ptr->hasChildGroup(*iter))
       {
         group_ptr = group_ptr->getGroup(*iter);
@@ -2326,7 +2326,7 @@ Group* Group::walkPath(std::string& path, bool create_groups_in_path)
         group_ptr = nullptr;
       }
     }
-    path = tokens.back();
+    path = path_parts.back();
   }
 
   return group_ptr;
@@ -2345,21 +2345,20 @@ const Group* Group::walkPath(std::string& path) const
 {
   const Group* group_ptr = this;
 
-  std::string::size_type pos = detail::find_exclusive(path, s_path_delimiter);
-  if(pos != std::string::npos)
+  // Split path into parts
+  std::vector<std::string> path_parts =
+    axom::Path(path, s_path_delimiter).parts();
+
+  if(path_parts.size() > 0)
   {
-    std::vector<std::string> tokens = detail::split(path, s_path_delimiter, pos);
-    std::vector<std::string>::iterator stop = tokens.end() - 1;
+    // Find stopping point (right before last part of path)
+    std::vector<std::string>::const_iterator stop = path_parts.end() - 1;
 
     // Navigate path down to desired Group
-    for(std::vector<std::string>::const_iterator iter = tokens.begin();
+    for(std::vector<std::string>::const_iterator iter = path_parts.begin();
         iter < stop;
         ++iter)
     {
-      SLIC_ASSERT_MSG(iter->size() > 0,
-                      SIDRE_GROUP_LOG_PREPEND << "Empty name in provided path '"
-                                              << path << "'.");
-
       if(group_ptr->hasChildGroup(*iter))
       {
         group_ptr = group_ptr->getGroup(*iter);
@@ -2370,7 +2369,7 @@ const Group* Group::walkPath(std::string& path) const
         iter = stop;
       }
     }
-    path = tokens.back();
+    path = path_parts.back();
   }
 
   return group_ptr;